# packages/contracts/generated/jarvis-api.openapi.json
{
  "components": {
    "schemas": {
      "AccountingPage": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AccountingRecord"
            },
            "title": "Items",
            "type": "array"
          },
          "next_after": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next After"
          }
        },
        "required": [
          "items"
        ],
        "title": "AccountingPage",
        "type": "object"
      },
      "AccountingRecord": {
        "additionalProperties": false,
        "properties": {
          "correlation_id": {
            "title": "Correlation Id",
            "type": "string"
          },
          "cost": {
            "$ref": "#/components/schemas/Cost"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "demo": {
            "default": false,
            "title": "Demo",
            "type": "boolean"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "latency_ms": {
            "title": "Latency Ms",
            "type": "integer"
          },
          "node_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Node Id"
          },
          "outcome": {
            "enum": [
              "completed",
              "failed",
              "unknown",
              "denied",
              "require_approval"
            ],
            "title": "Outcome",
            "type": "string"
          },
          "pricing": {
            "$ref": "#/components/schemas/Pricing-Output"
          },
          "profile_revision_id": {
            "format": "uuid",
            "title": "Profile Revision Id",
            "type": "string"
          },
          "project_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Id"
          },
          "provider_revision_id": {
            "format": "uuid",
            "title": "Provider Revision Id",
            "type": "string"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          },
          "route_revision_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Route Revision Id"
          },
          "run_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          },
          "task_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Task Id"
          },
          "usage": {
            "$ref": "#/components/schemas/Usage"
          }
        },
        "required": [
          "id",
          "profile_revision_id",
          "provider_revision_id",
          "correlation_id",
          "latency_ms",
          "usage",
          "pricing",
          "cost",
          "outcome",
          "created_at"
        ],
        "title": "AccountingRecord",
        "type": "object"
      },
      "ApiErrorDetail": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "maxLength": 120,
            "minLength": 1,
            "pattern": "^[a-z][a-z0-9_.-]*$",
            "title": "Code",
            "type": "string"
          },
          "details": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "title": "Details",
            "type": "object"
          },
          "message": {
            "maxLength": 1024,
            "minLength": 1,
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "maxLength": 64,
            "minLength": 8,
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "request_id"
        ],
        "title": "ApiErrorDetail",
        "type": "object"
      },
      "ApiErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "ApiErrorResponse",
        "type": "object"
      },
...
      "CandidateDecision": {
        "additionalProperties": false,
        "properties": {
          "eligible": {
            "title": "Eligible",
            "type": "boolean"
          },
          "profile_revision_id": {
            "format": "uuid",
            "title": "Profile Revision Id",
            "type": "string"
          },
          "reasons": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Reasons",
            "type": "array"
          }
        },
        "required": [
          "profile_revision_id",
          "eligible"
        ],
        "title": "CandidateDecision",
        "type": "object"
      },
...
      "Cost": {
        "additionalProperties": false,
        "properties": {
          "amount": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Amount"
          },
          "currency": {
            "default": "USD",
            "title": "Currency",
            "type": "string"
          },
          "status": {
            "default": "unknown",
            "enum": [
              "exact",
              "estimated",
              "unknown",
              "not_applicable"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "title": "Cost",
        "type": "object"
      },
...
      "FailureClass": {
        "enum": [
          "code.build_failure",
          "infrastructure.timeout",
          "unknown",
          "code.implementation_failure",
          "code.test_failure",
          "code.review_failure",
          "code.git_conflict",
          "infrastructure.worker_unavailable",
          "infrastructure.worker_transport",
          "infrastructure.service_unavailable",
          "provider.rate_limited",
          "provider.transient",
          "provider.contract_failure",
          "configuration.invalid",
          "security.policy_denied",
          "approval.rejected",
          "orchestration.runtime_error",
          "user.cancelled"
        ],
        "title": "FailureClass",
        "type": "string"
      },
      "JsonValue": {},
...
      "Pricing-Output": {
        "additionalProperties": false,
        "properties": {
          "cached_per_million": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cached Per Million"
          },
          "currency": {
            "default": "USD",
            "pattern": "^[A-Z]{3}$",
            "title": "Currency",
            "type": "string"
          },
          "effective_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effective At"
          },
          "input_per_million": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Input Per Million"
          },
          "output_per_million": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output Per Million"
          },
          "source": {
            "default": "Unspecified",
            "maxLength": 200,
            "title": "Source",
            "type": "string"
          },
          "status": {
            "default": "unknown",
            "enum": [
              "known",
              "unknown",
              "not_applicable"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "title": "Pricing",
        "type": "object"
      },
...
      "RoutePreviewRequest": {
        "additionalProperties": false,
        "properties": {
          "failed_profile_revision_ids": {
            "default": [],
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Failed Profile Revision Ids",
            "type": "array"
          },
          "failure_class": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FailureClass"
              },
              {
                "type": "null"
              }
            ]
          },
          "requirements": {
            "$ref": "#/components/schemas/RouteRequirements"
          },
          "route_revision_id": {
            "format": "uuid",
            "title": "Route Revision Id",
            "type": "string"
          }
        },
        "required": [
          "route_revision_id",
          "requirements"
        ],
        "title": "RoutePreviewRequest",
        "type": "object"
      },
      "RouteRequirements": {
        "additionalProperties": false,
        "properties": {
          "capabilities": {
            "default": [],
            "items": {
              "pattern": "^[a-z][a-z0-9_.-]{0,63}$",
              "type": "string"
            },
            "title": "Capabilities",
            "type": "array"
          },
          "data_classification": {
            "default": "public",
            "enum": [
              "public",
              "internal",
              "confidential",
              "restricted"
            ],
            "title": "Data Classification",
            "type": "string"
          },
          "input_tokens": {
            "default": 0,
            "maximum": 10000000.0,
            "minimum": 0.0,
            "title": "Input Tokens",
            "type": "integer"
          },
          "output_tokens": {
            "default": 1,
            "maximum": 1000000.0,
            "minimum": 1.0,
            "title": "Output Tokens",
            "type": "integer"
          },
          "purpose": {
            "pattern": "^[a-z][a-z0-9_.-]{0,63}$",
            "title": "Purpose",
            "type": "string"
          },
          "run_spend": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Spend"
          }
        },
        "required": [
          "purpose"
        ],
        "title": "RouteRequirements",
        "type": "object"
      },
      "RouteResolution": {
        "additionalProperties": false,
        "properties": {
          "candidates": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/CandidateDecision"
            },
            "title": "Candidates",
            "type": "array"
          },
          "decision": {
            "default": "deny",
            "enum": [
              "allow",
              "deny",
              "require_approval"
            ],
            "title": "Decision",
            "type": "string"
          },
          "demo": {
            "default": false,
            "title": "Demo",
            "type": "boolean"
          },
          "reasons": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Reasons",
            "type": "array"
          },
          "route_revision_id": {
            "format": "uuid",
            "title": "Route Revision Id",
            "type": "string"
          },
          "selected_profile_revision_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Selected Profile Revision Id"
          },
          "selected_provider_revision_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Selected Provider Revision Id"
          },
          "snapshot_hash": {
            "title": "Snapshot Hash",
            "type": "string"
          }
        },
        "required": [
          "route_revision_id",
          "snapshot_hash"
        ],
        "title": "RouteResolution",
        "type": "object"
      },
...
      "Usage": {
        "additionalProperties": false,
        "properties": {
          "cached_tokens": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cached Tokens"
          },
          "input_tokens": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Input Tokens"
          },
          "output_tokens": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output Tokens"
          },
          "provenance": {
            "default": "unknown",
            "enum": [
              "exact",
              "estimated",
              "unknown"
            ],
            "title": "Provenance",
            "type": "string"
          },
          "total_tokens": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Tokens"
          }
        },
        "title": "Usage",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Jarvis V1 API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/accounting": {
      "get": {
        "operationId": "list_accounting",
        "parameters": [
          {
            "in": "query",
            "name": "after",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "After"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountingPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Unsupported Media Type"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "List Accounting",
        "tags": [
          "routing"
        ]
      }
    },
...
    "/api/v1/routing/preview": {
      "post": {
        "operationId": "preview_route",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoutePreviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RouteResolution"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Unsupported Media Type"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Preview Route",
        "tags": [
          "routing"
        ]
      }
    },
...
  }
}